# Trivial performance suggestion for Linq `Where`