@EgorBo 2024-03-19
@EgorBot
```cs
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);

public class MyBench
{
    [Benchmark]
    public object Test() => new List<int>().Where(x => x > 0);
}
```