@stephentoub 2024-03-19
> why not just return this empty list and save allocating an extra ListWhereIterator.

Because the List could change in size before GetEnumerator is called and no longer be empty.