@alex-jitbit 2024-03-20
> Because the List could change in size before GetEnumerator is called and no longer be empty.

Gosh, I knew I missing something! This is so obvious. Apologies for wasting your time, will be more careful next time.